- fixed: publishingState filter is not processed probably in search API - remove legacy code
