+++ deploy/helm/internal-charts/indexer/templates/deployment.yaml
@@ -51,7 +51,6 @@ spec:
             "-Dregistry.baseUrl=http://registry-api",
             "-Dregistry.readOnlyBaseUrl=http://registry-api-read-only",
             "-DauthApi.baseUrl=http://authorization-api",
-            "-Dopa.baseUrl=http://authorization-api/v0/opa/",
             "-Dindexer.readSnapshots={{ .Values.readSnapshots }}",
             "-Dindexer.makeSnapshots={{ .Values.makeSnapshots }}",
             "-Dakka.loglevel={{ .Values.global.logLevel }}",
