+++ deploy/helm/internal-charts/search-api/templates/deployment.yaml
@@ -30,7 +30,6 @@ spec:
           - "-Dhttp.port=6102"
           - "-Dakka.loglevel={{ .Values.global.logLevel }}"
           - "-DauthApi.baseUrl=http://authorization-api"
-          - "-Dopa.baseUrl=http://authorization-api/v0/opa/"
           - "-DelasticSearch.serverUrl=elasticsearch://elasticsearch:9200"
 {{- if .Values.debug }}
           - "-DsearchApi.debug=true"
