+++ magda-registry-api/src/main/resources/application.conf
@@ -111,10 +111,6 @@ authApi {
     baseUrl = "http://localhost:6104"
 }
 
-opa {
-    baseUrl = ${authApi.baseUrl}"/v0/opa"
-}
-
 # whether validate received aspect data against aspect json schema
 validateJsonSchema = true
