+++ magda-scala-common/src/main/resources/common.conf
@@ -57,10 +57,6 @@ authApi {
     baseUrl = "http://localhost:6104"
 }
 
-opa {
-    baseUrl = ${authApi.baseUrl}"/v0/opa"
-}
-
 strings {
     unspecifiedWord = "Unspecified"
 }
